% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/kmeans.R
\docType{methods}
\name{ISDBSCAN}
\alias{ISDBSCAN}
\alias{ISDBSCAN,ANY-method}
\title{Mini-Batch k-means for large single cell sequencing data}
\usage{
ISDBSCAN(x, ...)

\S4method{ISDBSCAN}{ANY}(x, k, batch_size = blocksize(x))
}
\arguments{
\item{batch_size}{the size of the mini batches}

\item{clusters}{the number of clusters}

\item{num_init}{number of times the algorithm will be run with different
centroid seeds}

\item{max_iters}{the maximum number of clustering iterations}

\item{init_fraction}{percentage of data to use for the initialization
centroids (applies if initializer is \emph{kmeans++} ). Should be a float
number between 0.0 and 1.0.}

\item{initializer}{the method of initialization. One of \emph{kmeans++} and
\emph{random}. See details for more information}

\item{early_stop_iter}{continue that many iterations after calculation of the
best within-cluster-sum-of-squared-error}

\item{verbose}{either TRUE or FALSE, indicating whether progress is printed
during clustering}

\item{CENTROIDS}{a matrix of initial cluster centroids. The rows of the
CENTROIDS matrix should be equal to the number of clusters and the columns
should be equal to the columns of the data}

\item{tol}{a float number. If, in case of an iteration (iteration > 1 and
iteration < max_iters) 'tol' is greater than the squared norm of the
centroids, then kmeans has converged}

\item{seed}{integer value for random number generator (RNG)}
}
\value{
a list with the following attributes: centroids, WCSS_per_cluster,
 best_initialization, iters_per_initialization
}
\description{
This is an implementation of the mini-batch k-means algorithm of
  Sculley (2010) for large single cell sequencing data with the
  dimensionality reduction results as input in in the reducedDim() slot.
}
\details{
The implementation is largely based on the
  \code{\link[ClusterR]{MiniBatchKmeans}} function of the \code{ClusterR}
  package. The contribution of this package is to provide support for on-disk
  data representations such as HDF5, through the use of \code{DelayedMatrix}
  and \code{HDF5Matrix} objects, as well as for sparse data representation
  through the classes of the \code{Matrix} package.
  We also provide high-level methods for objects of class
  \code{SummarizedExperiment}, \code{SingleCellExperiment}, and
  \code{LinearEmbeddingMatrix}.

This function performs k-means clustering using mini batches.

\strong{kmeans++}: kmeans++ initialization. Reference :
http://theory.stanford.edu/~sergei/papers/kMeansPP-soda.pdf AND
http://stackoverflow.com/questions/5466323/how-exactly-does-k-means-work

\strong{random}: random selection of data rows as initial centroids
}
\references{
https://github.com/mlampros/ClusterR
}
